added ui

+++ internal/model/result.go
@@ -1,6 +1,6 @@
 package model
 
 type Result struct {
-	CorrectAnswers int     `json:"correctAnswers"`
-	Score          float64 `json:"score"`
+	CorrectAnswers int `json:"correctAnswers"`
+	Score          int `json:"score"`
 }
